Add cat to sentry modal.
We define a new set of constants which has paths, heights, and widths
for "flavor images". Then add one of those flavor images to the
sentry card select options. update teh selectOptions to accept the
new data (flavirImage) and update the component to render the flavor
image if an iamge was supplied.