Update Readme.md to be preciser
Oscar Wilde said beauty is terse.  And we all know that a design is complete not when there is nothing left to add, but simply when there is nothing left to takeaway. In that spirit, we modify the readme.md to more accurately attain true tonality.  After all, we who quarry simple stone must always be envisioning cathedrals.