Fix web socket type error.
The client sends an empty string to the server as a heartbeat.  That interval is initiated at the creation of the websocket on the client side.
The socket is stored in sockets.ws. The interval mistakenly references socket.send rather than sockets.ws.send
It is unaware of status of the connection on the server side.
This leads to trying to send messages on a socket that may not have a send available.
Fix this by checking that the websocket is connected before sending.