Remove card that user plays from hand.
Previously, when a user selected a card to play the client would send
the card name and the server would search for the card in that players
hand. And then play and remove it from their hand.

This works great, except when there are multiple cards with the same
name in the user's hand. If this is the case the server may select a
card of the same name, but in located in a different position in the
hand. While the effect will be the same, i.e. the cards are fungible
The users views these cards as different.
It's disorienting from their perspective if the card selected was not
played, but a different card was.

This was most apparent when playing treasures since they are more
likely to have duplicates in a single hand.  In these cases users
would complain that the game "reorders their hands".  No Bueno!

To fix this pass the index of the card played to the server.
Then use that index when removing the card from the array of
cards that represent the user's hand.

Cards in a players hands are represented as an array so passing the
index of the card is enough to uniquely identify the card.  We might
be able to get away with just passing the index, but who cares!